clean up project files
