--- app/src/main/java/com/apps005/magicstory/controller/MagicStoryApplication.java
@@ -1,12 +1,0 @@
-package com.apps005.magicstory.controller;
-
-
-import android.app.Application;
-
-public class MagicStoryApplication extends Application {
-    @Override
-    public void onCreate() {
-        super.onCreate();
-
-    }
-}
