+++ app/src/main/res/layout-land/activity_image_test.xml
@@ -11,8 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/light_white"
-        android:orientation="horizontal"
-        android:padding="10dp">
+        android:orientation="horizontal">
 
         <ImageView
             android:id="@+id/imageView"
@@ -22,17 +21,18 @@
             android:layout_marginTop="30dp"
             android:layout_marginBottom="30dp"
             android:layout_weight="1"
-            tools:srcCompat="@tools:sample/avatars"
-            android:contentDescription="@string/ai_generated_image_txt" />
+            android:contentDescription="@string/ai_generated_image_txt"
+            tools:srcCompat="@tools:sample/avatars" />
 
         <com.airbnb.lottie.LottieAnimationView
             android:id="@+id/animationView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:background="@color/anim_blue"
             android:visibility="gone"
             app:lottie_autoPlay="true"
             app:lottie_loop="true"
-            app:lottie_rawRes="@raw/writing_story" />
+            app:lottie_rawRes="@raw/anim_test" />
 
         <LinearLayout
             android:layout_width="match_parent"
