+++ app/src/main/res/layout/activity_image_test.xml
@@ -34,6 +34,7 @@
             android:id="@+id/animationView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:background="@color/anim_blue"
             android:visibility="gone"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="@+id/imageView"
@@ -41,7 +42,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:lottie_autoPlay="true"
             app:lottie_loop="true"
-            app:lottie_rawRes="@raw/writing_story" />
+            app:lottie_rawRes="@raw/anim_test" />
 
         <TextView
             android:id="@+id/read_story_statement"
